requirements.txt
Django==5.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-filter==24.1
drf-spectacular[sidecar]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
django-cors-headers==4.3.1
gunicorn==21.2.0
